move dvector into dmatrix
