+++ tests/vector_ops.rs
@@ -118,7 +118,7 @@ mod srvector_tests
 mod dvector_tests
 {
     use approx::assert_relative_eq;
-    use topohedral_linalg::dvector::{DVector, VecType};
+    use topohedral_linalg::dmatrix::{DVector, VecType};
     use topohedral_linalg::{FloatVectorOps, VectorOps};
 
     #[test]
